Move structEq to focal-atom and delete focal-utils
