+++ packages/focal-atom/src/index.ts
@@ -1,3 +1,4 @@
 export { Lens, Prism, Optic } from './lens/index'
 export { Atom, ReadOnlyAtom } from './atom/index'
 export { Option } from './utils'
+export { structEq } from './equals'
