+++ packages/focal-atom/tsconfig.build.es5.json
@@ -5,8 +5,5 @@
     "outDir": "dist/_esm5",
     "module": "commonjs",
     "target": "es2015",
-  },
-  "references": [
-    { "path": "../focal-utils/tsconfig.build.es5.json" }
-  ]
+  }
 }
