+++ packages/focal-atom/tsconfig.json
@@ -22,8 +22,5 @@
   "include": [
     "src/**/*.ts*",
     "test/**/*.ts*"
-  ],
-  "references": [
-    { "path": "../focal-utils/tsconfig.json" }
   ]
 }
