--- packages/focal-utils/package.json
@@ -1,93 +1,0 @@
-{
-  "name": "@grammarly/focal-utils",
-  "version": "0.8.1",
-  "description": "Focal common utils",
-  "main": "dist/_cjs/src/index.js",
-  "module": "dist/_esm5/src/index.js",
-  "es2015": "dist/_esm2015/src/index.js",
-  "types": "dist/_cjs/src/index.d.ts",
-  "files": [
-    "dist/_cjs/src/",
-    "dist/_esm5/src/",
-    "dist/_esm2015/src/"
-  ],
-  "scripts": {
-    "docs": "rm -rf ./docs && typedoc --out docs --theme minimal --ignoreCompilerErrors --tsconfig tsconfig.json",
-    "clean": "rm -rf ./dist",
-    "dev": "tsc -b -w tsconfig.build.cjs.json",
-    "build:cjs": "tsc -b tsconfig.build.cjs.json",
-    "build:es5": "tsc -b tsconfig.build.es5.json",
-    "build:es2015": "tsc -b tsconfig.build.es2015.json",
-    "build": "npm run clean && yarn build:cjs && yarn build:es5 && yarn build:es2015 && npm run lint",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "lint": "tslint --type-check --project tsconfig.json -c tslint.json"
-  },
-  "author": "Sergey Yavnyi <sergey.yavnyi@grammarly.com>",
-  "contributors": [
-    {
-      "name": "Sergey Yavnyi",
-      "email": "sergey.yavnyi@grammarly.com"
-    },
-    {
-      "name": "Valentyn Gaidylo",
-      "email": "valentyn.gaidylo@grammarly.com"
-    },
-    {
-      "name": "Igor Kononuchenko",
-      "email": "igor.kononuchenko@grammarly.com"
-    },
-    {
-      "name": "Alisa Mansurova",
-      "email": "alisa.mansurova@grammarly.com"
-    },
-    {
-      "name": "Eugene Chechuryn",
-      "email": "eugene.chechuryn@grammarly.com"
-    },
-    {
-      "name": "Yaroslav Voloshchuk",
-      "email": "yaroslav.voloshchuk@grammarly.com"
-    },
-    {
-      "name": "Sergey Rudenko",
-      "email": "sergey.rudenko@grammarly.com"
-    },
-    {
-      "name": "Gregory Shehet",
-      "email": "gregory.shehet@grammarly.com"
-    },
-    {
-      "name": "Anton Verinov",
-      "email": "anton@verinov.com"
-    },
-    {
-      "name": "Oleksii Levzhynskyi",
-      "email": "oleksii.levzhynskyi@gmail.com"
-    }
-  ],
-  "keywords": [
-    "frp",
-    "rx",
-    "rxjs",
-    "reactive",
-    "react",
-    "reactjs",
-    "lens"
-  ],
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/grammarly/focal.git"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@grammarly/tslint-config": "0.5.1",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^6.0.34",
-    "jest": "^24.1.0",
-    "ts-jest": "^23.10.5",
-    "tslint": "5.2.0",
-    "typescript": "3.3.3"
-  }
-}
