--- packages/focal-utils/src/index.ts
@@ -1,1 +1,0 @@
-export { equals as structEq } from './equals'
