--- packages/focal-utils/tsconfig.build.cjs.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": ["**/*.spec.ts*"],
-  "compilerOptions": {
-    "outDir": "dist/_cjs",
-    "module": "commonjs",
-    "target": "es5"
-  }
-}
