--- packages/focal-utils/tsconfig.build.es2015.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": ["**/*.spec.ts*"],
-  "compilerOptions": {
-    "outDir": "dist/_esm2015",
-    "module": "es2015",
-    "target": "esnext"
-  }
-}
