--- packages/focal-utils/tsconfig.build.es5.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": ["**/*.spec.ts*"],
-  "compilerOptions": {
-    "outDir": "dist/_esm5",
-    "module": "commonjs",
-    "target": "es2015",
-  }
-}
