+++ packages/focal/tsconfig.build.es2015.json
@@ -8,6 +8,5 @@
   },
   "references": [
     { "path": "../focal-atom/tsconfig.build.es2015.json" },
-    { "path": "../focal-utils/tsconfig.build.es2015.json" }
   ]
 }
